+++ libraries/matchers.rb
@@ -15,7 +15,7 @@ if defined?(ChefSpec)
   standard_actions = [:create, :create_if_missing, :delete]
   # Lists all the possible action verbs
   action_list = %w(create add delete remove set reset refresh update configure reconfigure edit none discover apply reapply activate
-                   stage new patch replace)
+                   stage new patch replace change)
 
   oneview_resources = {
     oneview_resource:                   standard_actions,
@@ -47,6 +47,7 @@ if defined?(ChefSpec)
                                          :replace_drive_enclosure],
     oneview_sas_interconnect:           [:reset, :hard_reset, :patch, :refresh, :set_uid_light, :set_power_state],
     oneview_sas_logical_interconnect_group: standard_actions,
+    oneview_scope:                      standard_actions + [:change_resource_assignments],
     oneview_server_hardware:            [:add_if_missing, :remove, :refresh, :set_power_state, :update_ilo_firmware, :patch],
     oneview_server_hardware_type:       [:edit, :remove],
     oneview_server_profile_template:    standard_actions + [:new_profile],