+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # 2.0.0
+Adds support to the following HPE OneView resources:
+  - Added oneview_scope resource (API 300 only)
+
 Adds support to API300 HPE Synergy Image Streamer resources:
   - image_streamer_plan_script